fix, test: 리뷰 관련 테스트 오류 수정
