+++ src/test/java/hanium/modic/backend/web/postReview/controller/PostReviewControllerIntegrationTest.java
@@ -101,73 +101,6 @@ class PostReviewControllerIntegrationTest extends BaseIntegrationTest {
 		assertThat(reviews.get(0).getDescription()).isEqualTo("좋은 리뷰입니다");
 	}
 
-	@Test
-	@DisplayName("TEST8: 리뷰 권한 확인 API - 권한 있음")
-	@WithCustomUser(email = "user1@test.com")
-	void canReviewPost_WithPermission_ShouldReturnTrue() throws Exception {
-		final UserEntity user = ContextHolderUtil.getCurrentUser();
-		// 다른 사용자가 만든 게시물 생성
-		final UserEntity postOwner = userEntityRepository.save(UserFactory.createMockUserWithoutId("postOwner"));
-		final PostEntity post = postEntityRepository.save(PostFactory.createMockPost(postOwner));
-
-		// AiImagePermission 생성 (해당 그림체를 사용한 이력 추가)
-		aiImagePermissionRepository.save(AiImagePermissionEntity.builder()
-			.userId(user.getId())
-			.postId(post.getId())
-			.remainingGenerations(10)
-			.isActive(true)
-			.build());
-
-		// when & then
-		mockMvc.perform(get("/api/post-reviews/can-review")
-				.param("postId", post.getId().toString()))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.isSuccess").value(true))
-			.andExpect(jsonPath("$.data.canReview").value(true));
-	}
-
-	@Test
-	@DisplayName("TEST9: 리뷰 권한 확인 API - 권한 없음")
-	@WithCustomUser(email = "user2@test.com")
-	void canReviewPost_WithoutPermission_ShouldReturnFalse() throws Exception {
-		final UserEntity user = ContextHolderUtil.getCurrentUser();
-		// 다른 사용자가 만든 게시물 생성
-		final UserEntity postOwner = userEntityRepository.save(UserFactory.createMockUserWithoutId("postOwner"));
-		final PostEntity post = postEntityRepository.save(PostFactory.createMockPost(postOwner));
-
-		// AiImagePermission 생성하지 않음 (해당 그림체를 사용한 이력 없음)
-
-		// when & then
-		mockMvc.perform(get("/api/post-reviews/can-review")
-				.param("postId", post.getId().toString()))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.isSuccess").value(true))
-			.andExpect(jsonPath("$.data.canReview").value(false));
-	}
-
-	@Test
-	@DisplayName("TEST10: 리뷰 권한 확인 API - 자신의 게시물")
-	@WithCustomUser(email = "user3@test.com")
-	void canReviewPost_OwnPost_ShouldReturnFalse() throws Exception {
-		final UserEntity user = ContextHolderUtil.getCurrentUser();
-		// 자신의 게시물 생성
-		final PostEntity post = postEntityRepository.save(PostFactory.createMockPost(user));
-
-		// AiImagePermission 생성 (해당 그림체를 사용한 이력 추가)
-		aiImagePermissionRepository.save(AiImagePermissionEntity.builder()
-			.userId(user.getId())
-			.postId(post.getId())
-			.remainingGenerations(10)
-			.isActive(true)
-			.build());
-
-		// when & then
-		mockMvc.perform(get("/api/post-reviews/can-review")
-				.param("postId", post.getId().toString()))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.isSuccess").value(true))
-			.andExpect(jsonPath("$.data.canReview").value(false));
-	}
 
 	@Test
 	@DisplayName("TEST2: 리뷰 수정 성공 - 본인")
